Move `GetInputs` and `GetOutputs` from `TransactionHistoryBuilder` to `SmartTransactionExtensions`.
